planner_cspace: fix flaky action tests
