+++ planner_cspace/test/include/planner_cspace/action_test_base.h
@@ -1,0 +1,94 @@
+/*
+ * Copyright (c) 2018-2023, the neonavigation authors
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ *
+ *     * Redistributions of source code must retain the above copyright
+ *       notice, this list of conditions and the following disclaimer.
+ *     * Redistributions in binary form must reproduce the above copyright
+ *       notice, this list of conditions and the following disclaimer in the
+ *       documentation and/or other materials provided with the distribution.
+ *     * Neither the name of the copyright holder nor the names of its
+ *       contributors may be used to endorse or promote products derived from
+ *       this software without specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+ * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+ * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+ * ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE
+ * LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
+ * CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
+ * SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
+ * INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
+ * CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
+ * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+ * POSSIBILITY OF SUCH DAMAGE.
+ */
+
+#ifndef PLANNER_CSPACE_ACTION_TEST_BASE_H
+#define PLANNER_CSPACE_ACTION_TEST_BASE_H
+
+#include <memory>
+#include <string>
+
+#include <gtest/gtest.h>
+
+#include <ros/ros.h>
+
+#include <actionlib/client/simple_action_client.h>
+#include <move_base_msgs/MoveBaseAction.h>
+#include <planner_cspace_msgs/PlannerStatus.h>
+#include <tf2_ros/transform_listener.h>
+
+const char ACTION_TOPIC_MOVE_BASE[] = "/move_base";
+const char ACTION_TOPIC_TOLERANT_MOVE[] = "/tolerant_move";
+
+template <typename ACTION, char const* TOPIC>
+class ActionTestBase : public ::testing::Test
+{
+public:
+  ActionTestBase()
+    : tfl_(tfbuf_)
+  {
+    move_base_ = std::make_shared<ActionClient>(TOPIC);
+    status_sub_ = node_.subscribe(
+        "/planner_3d/status", 10, &ActionTestBase::cbStatus, this);
+    if (!move_base_->waitForServer(ros::Duration(30.0)))
+    {
+      ROS_ERROR("Failed to connect move_base action");
+      exit(EXIT_FAILURE);
+    }
+  }
+  ~ActionTestBase()
+  {
+  }
+
+protected:
+  using ActionClient = actionlib::SimpleActionClient<ACTION>;
+  using ActionClientPtr = std::shared_ptr<ActionClient>;
+
+  void cbStatus(const planner_cspace_msgs::PlannerStatus::ConstPtr& msg)
+  {
+    planner_status_ = msg;
+  }
+  std::string statusString() const
+  {
+    if (!planner_status_)
+    {
+      return "(no status)";
+    }
+    return "(status: " + std::to_string(planner_status_->status) +
+           ", error: " + std::to_string(planner_status_->error) + ")";
+  }
+
+  ros::NodeHandle node_;
+  ros::Subscriber status_sub_;
+  ActionClientPtr move_base_;
+  planner_cspace_msgs::PlannerStatus::ConstPtr planner_status_;
+  tf2_ros::Buffer tfbuf_;
+  tf2_ros::TransformListener tfl_;
+};
+
+#endif  // PLANNER_CSPACE_ACTION_TEST_BASE_H
